requirements.txt
Flask==3.0.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
marshmallow==3.21.3
Flask-Marshmallow==1.2.1
marshmallow-sqlalchemy==1.1.0
python-dotenv==1.0.1
gunicorn==22.0.0

# Dev / test
pytest==8.2.2
pytest-flask==1.3.0
